+++ src/drivers/JsonApi.ts
@@ -1,6 +1,6 @@
 
 import { basepath } from '../internal/utility';
-import { IDriver, ConnectionState, Identity, Permission } from '../types';
+import { IDriver, ConnectionState, Identity, DriverResponse } from '../types';
 
 /** Payload from a JSON:API IdM endpoint */
 interface IdentityJsonApiResponse {
@@ -10,12 +10,12 @@ interface IdentityJsonApiResponse {
         username: string;
         email: string;
         name: string;
-            
+
         /**
          * Active application permissions
          */
-        permissions: Permission[];
-            
+        permissions: string[];
+
         emulation: {
             active: boolean;
             allowed: boolean;
@@ -36,8 +36,8 @@ class JsonApiDriver implements IDriver
         this.emulateEndpoint = emulateEndpoint;
     }
 
-    public async refreshIdentity(): Promise<[ConnectionState, Identity]> {
-        // Grab updated identity information. Non-200 response 
+    public async refreshIdentity(): Promise<DriverResponse> {
+        // Grab updated identity information. Non-200 response
         // is the assumption that we're no longer authenticated
         const res = await fetch(this.endpoint, {
             cache: 'no-cache',
@@ -48,32 +48,45 @@ class JsonApiDriver implements IDriver
             },
         });
 
-        const { data } = await res.json();
+        // Try to parse out user data from the response payload.
+        try {
+            const { data } = await res.json();
 
-        // Ensure the payload is not malformed
-        if (typeof data.attributes === 'undefined' || data.type !== 'User') {
-            throw new Error(`Malformed identity API response: ${JSON.stringify(data)}`);
+            // Ensure the payload is not malformed
+            if (typeof data.attributes === 'undefined' || data.type !== 'User') {
+                throw new Error();
+            }
+
+            const jsonApiIdentity = data as IdentityJsonApiResponse;
+            const user: Identity = {
+                id: jsonApiIdentity.id,
+
+                // Make extractions explicit, otherwise developers may add
+                // extra stuff that won't be supported in the future AWS-lands.
+                // Making migration that much harder.
+                name: jsonApiIdentity.attributes.name,
+                username: jsonApiIdentity.attributes.username,
+                email: jsonApiIdentity.attributes.email,
+                permissions: jsonApiIdentity.attributes.permissions,
+                emulation: jsonApiIdentity.attributes.emulation,
+            };
+
+            // If it parsed correctly, pass the identity forward
+            return {
+                state: ConnectionState.LOGGED_IN,
+                user,
+                error: undefined
+            };
+        } catch (e) {
+            console.error('[auth]- JSON:API Driver Error:', e);
+
+            // Any errors will be caught as a parsing issue from the API.
+            return {
+                state: ConnectionState.API_ERROR,
+                user: undefined,
+                error: 'The server did not provide valid user information'
+            }
         }
-
-        const jsonApiIdentity = data as IdentityJsonApiResponse;
-        const identity: Identity = {
-            id: jsonApiIdentity.id,
-            // ...jsonApiIdentity.attributes
-            
-            // Make extractions explicit, otherwise developers may add  
-            // extra stuff that won't be supported in the future AWS-lands.
-            // Making migration that much harder. 
-            name: jsonApiIdentity.attributes.name,
-            username: jsonApiIdentity.attributes.username,
-            email: jsonApiIdentity.attributes.email,
-            permissions: jsonApiIdentity.attributes.permissions,
-            emulation: jsonApiIdentity.attributes.emulation,
-
-            // Policies are currently not supported by JSON:API.
-            policies: [] 
-        };
-
-        return [ConnectionState.LOGGED_IN, identity];
     }
 
     public async emulate(id: string): Promise<void> {
@@ -85,7 +98,7 @@ class JsonApiDriver implements IDriver
             }
         });
     }
-    
+
     public async clearEmulation(): Promise<void> {
         await fetch(this.emulateEndpoint, {
             method: 'DELETE',