+++ src/index.ts
@@ -1,27 +1,31 @@
 
 import AuthProvider from './components/AuthProvider';
-import AuthenticationMonitor from './components/AuthenticationMonitor';
+import ReauthenticateModal from './components/ReauthenticateModal';
 import Profile from './components/Profile';
+import SimpleSignOutButton from './components/SimpleSignOutButton';
 import Can from './components/Can';
-import AuthContext from './context/AuthContext';
 import useIdentity from './hooks/useIdentity';
 import useEmulation from './hooks/useEmulation';
+import GraphQL from './drivers/GraphQL';
+import JsonApi from './drivers/JsonApi';
 import { IDriver, IHasPolicies } from './types';
 
 export {
     // Components
     AuthProvider,
-    AuthenticationMonitor,
+    ReauthenticateModal,
     Profile,
+    SimpleSignOutButton,
     Can,
-
-    // Context
-    AuthContext,
 
     // Hooks
     useIdentity,
     useEmulation,
-};
+
+    // Drivers
+    GraphQL,
+    JsonApi,
+}
 
 export type {
     // Interfaces